feat: initialize product status tables

+++ src/Core/Modules/Product/productRespository.cs
@@ -1,5 +1,7 @@
 ﻿using e_commerce_server.src.Core.Database.Data;
 using e_commerce_server.src.Core.Modules.Product.Dto;
+using e_commerce_server.src.Core.Modules.Product.Service;
+using e_commerce_server.Src.Core.Common.Enum;
 using e_commerce_server.Src.Core.Database.Data;
 using e_commerce_server.Src.Core.Modules.User;
 using e_commerce_server.Src.Packages.HttpException;
@@ -27,13 +29,13 @@ namespace e_commerce_server.src.Core.Modules.Product
                 throw new InternalException(ex.Message);
             }
         }
-        public List<object> GetProductByPage(DbSet<ProductData> products,int page, int pageSize)
+        public List<object> GetProductsByPage(DbSet<ProductData> products, int page)
         {
             try
             {
                 return products
                     .Skip((page -1) * 10)
-                    .Take(pageSize)
+                    .Take(PageSizeEnum.PAGE_SIZE)
                     .Select(
                     p => new
                     {
@@ -44,10 +46,9 @@ namespace e_commerce_server.src.Core.Modules.Product
                         p.description,
                         p.created_at,
                         p.updated_at,
-                        p.product_status,
+                        product_status = p.product_status.status,
                         thumbnails = _context.Thumbnails.Where(cond => cond.product_id == p.id).Select(data => data.thumbnail_url).ToList(),
                         p.user_id,
-                        p.category_id,
                         category = p.category.name,
                         address = _context.Districts
                             .Where(a => a.id == p.user.district_id)
@@ -55,18 +56,23 @@ namespace e_commerce_server.src.Core.Modules.Product
                     }
                 ).Cast<object>().ToList();
             } catch (Exception ex)
-                {
-                    throw new InternalException(ex.Message);
-                }
+            {
+                throw new InternalException(ex.Message);
+            }
         }
-        public ProductData UpdateProduct(int productID, ProductDto productDto)
+        public ProductData UpdateProduct(int productId, ProductDto productDto)
         {
             try
             {
-                var product = _context.Products.SingleOrDefault(p => p.id == productID);
+                var product = _context.Products.SingleOrDefault(p => p.id == productId);
 
-                //delete list old thumbnail of product
-                var thumbnailsToDelete = _context.Thumbnails.Where(cond => cond.product_id == productID).ToList();
+                if (product == null)
+                {
+                    throw new BadRequestException(ProductEnum.PRODUCT_NOT_FOUND);
+                }
+
+                //delete old product's thumbnails
+                var thumbnailsToDelete = _context.Thumbnails.Where(cond => cond.product_id == productId).ToList();
 
                 _context.Thumbnails.RemoveRange(thumbnailsToDelete);
 
@@ -77,7 +83,7 @@ namespace e_commerce_server.src.Core.Modules.Product
                     ThumbnailData item = new ThumbnailData
                     {
                         thumbnail_url = thumbnail,
-                        product_id = productID,
+                        product_id = productId,
                     };
 
                     thumbnails.Add(item);
@@ -96,8 +102,8 @@ namespace e_commerce_server.src.Core.Modules.Product
 
                 return product;
             } catch (Exception ex)
-                {
-                    throw new InternalException(ex.Message);
+            {
+                throw new InternalException(ex.Message);
             }
         }
         public ProductData AddProduct(ProductDto productDto,int idUser)
@@ -112,7 +118,7 @@ namespace e_commerce_server.src.Core.Modules.Product
                     discount = productDto.discount,
                     created_at = DateTime.Now,
                     updated_at = DateTime.Now,
-                    product_status = productDto.product_status,
+                    status_id = productDto.status_id,
                     user_id = idUser,
                     category_id = productDto.category_id,
                 };
@@ -140,34 +146,18 @@ namespace e_commerce_server.src.Core.Modules.Product
                 throw new InternalException(e.Message);
             }
         }
-        public List<string> GetThumbnailsByProductId(int id)
-        {
-            try
-            {
-                return _context.Thumbnails.Where(cond => cond.product_id == id).Select(data => data.thumbnail_url).ToList();
-            } catch (Exception e)
-            {
-                throw new InternalException(e.Message);
-            }
-        }
         public ProductDto? GetProductById(int id)
         {
             try
             {
                 var product = _context.Products.FirstOrDefault(p => p.id == id);
 
+                var user = userRepository.GetById(product.user_id);
+
                 if (product == null) 
                 {
                     return null;
                 }
-
-                var thumbnails = GetThumbnailsByProductId(id);
-
-                var user = userRepository.FindById(product.user_id);
-
-                var address = _context.Districts
-                        .Where(a => a.id == user.district_id)
-                        .Select(ct => ct.city.name).SingleOrDefault();
 
                 return new ProductDto
                 {
@@ -178,11 +168,13 @@ namespace e_commerce_server.src.Core.Modules.Product
                     description = product.description,
                     created_at = product.created_at,
                     updated_at = product.updated_at,
-                    product_status = product.product_status,
-                    thumbnails = thumbnails,
-                    address = address,
-                    user_id = product.user.id,
+                    status_id = product.status_id,
+                    thumbnails = _context.Thumbnails.Where(cond => cond.product_id == product.id).Select(data => data.thumbnail_url).ToList(),
+                    user_id = product.user_id,
                     category_id = product.category_id,
+                    address = _context.Districts
+                        .Where(a => a.id == product.user.district_id)
+                        .Select(ct => ct.city.name).SingleOrDefault()
                 };
             }
             catch (Exception ex)
